+++ CHANGELOG.rst
@@ -16,6 +16,7 @@
 - Make ``logger.catch()`` usable as an asynchronous context manager (`#1084 <https://github.com/Delgan/loguru/issues/1084>`_).
 - Make ``logger.catch()`` compatible with asynchronous generators (`#1302 <https://github.com/Delgan/loguru/issues/1302>`_).
 - Improve feedback for invalid format keys in logger format strings (`#1450 <https://github.com/Delgan/loguru/issues/1450>`_, thanks `@Krishnachaitanyakc <https://github.com/Krishnachaitanyakc>`_).
+- Make callable log ``format`` append the same line ending and ``{exception}`` suffix as string formats (`#1454 <https://github.com/Delgan/loguru/issues/1454>`_).
 
 `0.7.3`_ (2024-12-06)
 =====================